+++ RELEASENOTES.md
@@ -11,6 +11,8 @@
         ([#10057](https://github.com/google/ExoPlayer/issues/10057)).
     *   Limit parallel download removals to 1 to avoid excessive thread creation
         ([#10458](https://github.com/google/ExoPlayer/issues/10458)).
+    *   Discard back buffer before playback gets stuck due to insufficient
+        available memory.
 *   Downloads:
     *   Fix potential infinite loop in `ProgressiveDownloader` caused by
         simultaneous download and playback with the same `PriorityTaskManager`